refactor(analytics): extract AnalyticsKpiCards component (KPI cards + CV breakdown)
